Move namespace-usage assertions into the Namespace package
Namespace wiring isn't a variant of the feature that happens to own it, so per the
agreed strategy it belongs in the Namespace package rather than as extra methods on
instantiation test classes.

Relocated unchanged (only renamed where the old name no longer needed to
disambiguate, since the module now supplies that context):
- Concurrent.py::ConcurrentBlockStatements::test_NestedBlock_ConnectsNamespaceOnParentAssignment
    -> Namespace/Statements.py::BlockStatements (same name)
- Concurrent.py::IfGenerateStatements::test_IfElsifElse_ConnectsNamespaces
    -> Namespace/Statements.py::IfGenerateStatements::test_AllBranchesConnectToTheArchitecture
- Concurrent.py::CaseGenerateStatements::test_ConnectsNamespaceOnParentAssignment
    -> Namespace/Statements.py::CaseGenerateStatements::test_CaseConnectsToTheArchitecture
- Concurrent.py::ForGenerateStatements::test_ConnectsNamespaceOnParentAssignment
    -> Namespace/Statements.py::ForGenerateStatements::test_ConnectsToTheArchitecture
- DesignUnit.py::Architectures::test_IndexDeclaredItems_AlsoPopulatesNamespace
    -> Namespace/DesignUnits.py::Architectures (same name)

The note about namespace-connection happening in the `Parent` setter override
rather than the constructor moved up to the Statements module docstring, since it
applies to every test in it. Test count is unchanged at 387.
