+++ tests/unit/Namespace/DesignUnits.py
@@ -1,0 +1,57 @@
+# ==================================================================================================================== #
+#             __     ___   _ ____  _     __  __           _      _                                                     #
+#   _ __  _   \ \   / / | | |  _ \| |   |  \/  | ___   __| | ___| |                                                    #
+#  | '_ \| | | \ \ / /| |_| | | | | |   | |\/| |/ _ \ / _` |/ _ \ |                                                    #
+#  | |_) | |_| |\ V / |  _  | |_| | |___| |  | | (_) | (_| |  __/ |                                                    #
+#  | .__/ \__, | \_/  |_| |_|____/|_____|_|  |_|\___/ \__,_|\___|_|                                                    #
+#  |_|    |___/                                                                                                        #
+# ==================================================================================================================== #
+# Authors:                                                                                                             #
+#   Patrick Lehmann                                                                                                    #
+#                                                                                                                      #
+# License:                                                                                                             #
+# ==================================================================================================================== #
+# Copyright 2026-2026 Patrick Lehmann - Boetzingen, Germany                                                            #
+#                                                                                                                      #
+# Licensed under the Apache License, Version 2.0 (the "License");                                                      #
+# you may not use this file except in compliance with the License.                                                     #
+# You may obtain a copy of the License at                                                                              #
+#                                                                                                                      #
+#   http://www.apache.org/licenses/LICENSE-2.0                                                                         #
+#                                                                                                                      #
+# Unless required by applicable law or agreed to in writing, software                                                  #
+# distributed under the License is distributed on an "AS IS" BASIS,                                                    #
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.                                             #
+# See the License for the specific language governing permissions and                                                  #
+# limitations under the License.                                                                                       #
+#                                                                                                                      #
+# SPDX-License-Identifier: Apache-2.0                                                                                  #
+# ==================================================================================================================== #
+#
+"""Namespaces owned by design units, and how ``IndexDeclaredItems`` populates them."""
+from unittest import TestCase
+
+from pyVHDLModel.DesignUnit import Architecture
+from pyVHDLModel.Name       import SimpleName
+from pyVHDLModel.Object     import Constant
+from pyVHDLModel.Symbol     import EntitySymbol, SimpleSubtypeSymbol
+
+
+if __name__ == "__main__":  # pragma: no cover
+	print("ERROR: you called a testcase declaration file as an executable module.")
+	print("Use: 'python -m unitest <testcase module>'")
+	exit(1)
+
+
+def _entitySymbol(name: str = "e") -> EntitySymbol:
+	return EntitySymbol(SimpleName(name))
+
+
+class Architectures(TestCase):
+	def test_IndexDeclaredItems_AlsoPopulatesNamespace(self) -> None:
+		constant = Constant(["C"], SimpleSubtypeSymbol(SimpleName("natural")))
+		architecture = Architecture("rtl", _entitySymbol(), declaredItems=[constant])
+		architecture.IndexDeclaredItems()
+
+		# Keyed by the normalized identifier, which is what the Find* methods look up.
+		self.assertIs(constant, architecture._namespace.Elements()["c"])
